+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and
 
 ## [Unreleased]
 
-## [0.2.1] - 2026-07-13
+## [0.2.1] - 2026-07-14
 
 ### Fixed